refactor(tests): register AWS matcher alongside existing matchers
