+++ packages/kafka/tests/tsconfig.json
@@ -5,5 +5,5 @@
     "noEmit": true,
     "resolveJsonModule": true
   },
-  "include": ["../../testing/src/setupEnv.ts", "../../testing/src/awsSdkMatchers.ts", "../src/**/*", "./**/*"]
+  "include": ["../../testing/src/setupEnv.ts", "../src/**/*", "./**/*"]
 }
