+++ packages/metrics/tests/tsconfig.json
@@ -4,5 +4,5 @@
     "rootDir": "../../",
     "noEmit": true
   },
-  "include": ["../../testing/src/setupEnv.ts", "../../testing/src/awsSdkMatchers.ts", "../src/**/*", "./**/*"]
+  "include": ["../../testing/src/setupEnv.ts", "../src/**/*", "./**/*"]
 }
