+++ packages/testing/src/setupEnv.ts
@@ -1,7 +1,6 @@
-import { expect, vi } from 'vitest';
-import { type AwsSdkMatchers, toReceiveCommandWith } from './awsSdkMatchers.js';
-
-expect.extend({ toReceiveCommandWith });
+import type { MetadataBearer } from '@smithy/types';
+import type { AwsCommand, AwsStub } from 'aws-sdk-client-mock';
+import { expect, type MatcherResult, vi } from 'vitest';
 
 // Mock console methods to prevent output during tests
 vi.spyOn(console, 'error').mockReturnValue();
@@ -11,6 +10,33 @@ vi.spyOn(console, 'info').mockReturnValue();
 vi.spyOn(console, 'log').mockReturnValue();
 
 expect.extend({
+  /**
+   * Matches recorded AWS SDK command inputs using Vitest's partial object matching.
+   *
+   * @param received - The mocked AWS SDK client
+   * @param command - The AWS SDK command constructor
+   * @param expected - The expected subset of the command input
+   */
+  toReceiveCommandWith<Input extends object, Output extends MetadataBearer>(
+    received: AwsStub<Input, Output, unknown>,
+    command: new (input: Input) => AwsCommand<Input, Output>,
+    expected: Partial<Input>
+  ): MatcherResult {
+    const inputs = received
+      .commandCalls(command)
+      .map((call) => call.args[0].input);
+    const pass = inputs.some((input) =>
+      this.equals(input, expect.objectContaining<object>(expected))
+    );
+
+    return {
+      pass,
+      message: () =>
+        `Expected ${received.clientName()} ${command.name} ${this.isNot ? 'not ' : ''}to receive input containing ${this.utils.printExpected(expected)}\nReceived inputs (call count: ${inputs.length}): ${this.utils.printReceived(inputs)}`,
+      actual: inputs,
+      expected,
+    };
+  },
   toHaveLogged(received, expected) {
     const calls = received.mock.calls;
     const messages = new Array(calls.length);
@@ -228,6 +254,22 @@ expect.addEqualityTesters([
     return aOk === bOk ? undefined : false;
   },
 ]);
+
+/**
+ * Describes the AWS SDK matchers available in test assertions.
+ */
+interface AwsSdkMatchers {
+  /**
+   * Asserts that at least one call to the command contains the expected input.
+   *
+   * @param command - The AWS SDK command constructor
+   * @param expected - The expected subset of the command input
+   */
+  toReceiveCommandWith<Input extends object, Output extends MetadataBearer>(
+    command: new (input: Input) => AwsCommand<Input, Output>,
+    expected: Partial<NoInfer<Input>>
+  ): void;
+}
 
 declare module 'vitest' {
   // biome-ignore lint/suspicious/noExplicitAny: vitest typings expect an any type
